requirements.txt
build
cffi
opencv-python
numpy
pygame
leapc-python-api